+++ src/vertagus/core/project.py
@@ -1,8 +1,9 @@
 import typing as T
 from logging import getLogger
+from copy import copy
 
 from vertagus.core.manifest_base import ManifestBase
-from vertagus.core.rule_bases import SingleVersionRule, VersionComparisonRule
+from vertagus.core.rule_bases import SingleVersionRuleProtocol, VersionComparisonRule
 from vertagus.rules.comparison.library import ManifestsComparisonRule
 from vertagus.core.tag_base import AliasBase
 from vertagus.core.bumper_base import BumperBase
@@ -24,7 +25,7 @@ class Project(Package):
     def __init__(
         self,
         manifests: list[ManifestBase],
-        current_version_rules: list[T.Type[SingleVersionRule]],
+        current_version_rules: list[SingleVersionRuleProtocol],
         version_increment_rules: list[VersionComparisonRule],
         manifest_versions_comparison_rules: list[ManifestsComparisonRule],
         stages: T.Optional[list[Stage]] = None,
@@ -138,7 +139,7 @@ class Project(Package):
             manifests.extend(stage.manifests)
         return list(dict.fromkeys(manifests).keys())
 
-    def _get_current_version_rules(self, stage_name=None) -> list[SingleVersionRule]:
+    def _get_current_version_rules(self, stage_name=None) -> list[SingleVersionRuleProtocol]:
         rules = self._current_version_rules.copy()
         if stage_name:
             stage = self._get_stage(stage_name)
@@ -153,7 +154,7 @@ class Project(Package):
         return list(dict.fromkeys(rules).keys())
 
     def _get_manifest_versions_comparison_rules(self, stage_name=None) -> list[ManifestsComparisonRule]:
-        rules = self._manifest_versions_comparison_rules.copy()
+        rules = list(copy(self._manifest_versions_comparison_rules))
         if stage_name:
             stage = self._get_stage(stage_name)
             rules.extend(stage.manifest_versions_comparison_rules)